requirements.txt
# Core
neo4j>=5.20.0
python-dotenv>=1.0.0

# Scientific computing
numpy>=1.26.0
scipy>=1.12.0
pandas>=2.2.0

# Graph analysis
networkx>=3.2.0

# Topological data analysis
ripser>=0.6.4
persim>=0.3.2
scikit-learn>=1.4.0

# Formal concept analysis
concepts>=0.9.2

# Visualisation
matplotlib>=3.8.0
seaborn>=0.13.0
pyvis>=0.3.2

# Power-law fitting
powerlaw>=1.5

# Data acquisition (DAQ) pipeline
requests>=2.31.0
tqdm>=4.66.0

# Ontology generation
rdflib>=7.0.0